backend/requirements.txt
fastapi
uvicorn
google-generativeai
httpx
python-dotenv
pydantic
resend
